examples/tau-bench: rename tool_parser.py → vllm_tool_parser.py
Adds a 'vllm' prefix so the filename signals where the parser was
ported from (vLLM's Hermes2ProToolParser) and avoids the ambiguous
generic name.

Updated callers:
  - openai_tool_adapter.py: 'from tool_parser' → 'from vllm_tool_parser'
  - trainable_agents.py docstring: 'tool_parser.parse_tools' → 'vllm_tool_parser.parse_tools'
