+++ examples/tau-bench/vllm_tool_parser.py
